+++ R/score_F1.R
@@ -1,9 +1,10 @@
 #' @title F1 Score
 #'
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values (y_hat) to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
-#'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -20,12 +21,13 @@
 #'
 #'
 #' @export
-
-
-score_f1 <- function(object){
+score_f1 <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
 
   ret <- (2 * (conf$TP / (conf$TP + conf$FP)) * (conf$TP / (conf$TP + conf$FN))) /
     (conf$TP / (conf$TP + conf$FN) + conf$TP / (conf$TP + conf$FP))
@@ -41,10 +43,11 @@ score_f1 <- function(object){
 
 #' @title One Minus F1 Score
 #'
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values (y_hat) to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
-#'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -61,12 +64,13 @@ score_f1 <- function(object){
 #'
 #'
 #' @export
-
-
-score_one_minus_f1 <- function(object){
+score_one_minus_f1 <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
 
   ret <- (2 * (conf$TP / (conf$TP + conf$FP)) * (conf$TP / (conf$TP + conf$FN))) /
     (conf$TP / (conf$TP + conf$FN) + conf$TP / (conf$TP + conf$FP))