+++ R/score_acc.R
@@ -1,8 +1,10 @@
 #' @title Accuracy
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values (y_hat) to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -19,12 +21,14 @@
 #'
 #'
 #' @export
-
-
-score_acc <- function(object){
+#' @rdname score_acc
+score_acc <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
   ret <- (conf$TP + conf$TN) / (conf$TP + conf$FP + conf$TN + conf$FN)
 
   acc_results <- list(
@@ -39,9 +43,11 @@ score_acc <- function(object){
 
 #' @title One minus accuracy
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -58,12 +64,14 @@ score_acc <- function(object){
 #'
 #'
 #' @export
-
-
-score_one_minus_acc <- function(object){
+#' @rdname score_one_minus_acc
+score_one_minus_acc <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
   ret <- 1 - (conf$TP + conf$TN) / (conf$TP + conf$FP + conf$TN + conf$FN)
 
   acc_results <- list(