+++ R/score_specificity.R
@@ -1,8 +1,10 @@
 #' @title Specificity
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values (y_hat) to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -21,10 +23,13 @@
 #' @export
 
 
-score_specificity <- function(object){
+score_specificity <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
   ret <- conf$TN / (conf$TN + conf$FP)
 
   specificity_results <- list(
@@ -39,9 +44,11 @@ score_specificity <- function(object){
 
 #' @title One minus specificity
 #'
-#' @param object An object of class 'explainer' created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param object An object of class \code{explainer} created with function \code{\link[DALEX]{explain}} from the DALEX package.
+#' @param cutoff Treshold value, which divides model predicted values (y_hat) to calculate confusion matrix. By default it's \code{0.5}.
+#' @param data New data that will be used to calcuate the score. Pass \code{NULL} if you want to use \code{data} from \code{object}.
 #'
-#' @return An object of class 'auditor_score'.
+#' @return An object of class \code{auditor_score}.
 #'
 #' @examples
 #' titanic <- na.omit(DALEX::titanic)
@@ -60,10 +67,13 @@ score_specificity <- function(object){
 #' @export
 
 
-score_one_minus_specificity <- function(object){
+score_one_minus_specificity <- function(object, cutoff = 0.5, data = NULL) {
   if(!("explainer" %in% class(object))) stop("The function requires an object created with explain() function from the DALEX package.")
 
-  conf <- confusionmatrix(object)
+  # inject new data to the explainer
+  if (!is.null(data)) object$data <- data
+
+  conf <- confusionmatrix(object, cutoff)
   ret <- 1 - conf$TN / (conf$TN + conf$FP)
 
   specificity_results <- list(